Andrew patrick

+++ feed/src/main/java/com/revature/feed/controller/LikeController.java
@@ -1,5 +1,7 @@
 package com.revature.feed.controller;
 
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.revature.feed.config.JwtUtility;
 import com.revature.feed.models.Like;
 import com.revature.feed.models.Response;
 import com.revature.feed.services.LikeService;
@@ -8,6 +10,7 @@ import org.springframework.core.env.Environment;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController("likeController")
 @RequestMapping(value= "like")
@@ -20,10 +23,19 @@ public class LikeController {
     @Autowired
     private Environment environment;
 
+    @Autowired
+    JwtUtility jwtUtility;
+
     //Create a Like
     //Angular send message we receive
     @PostMapping
-    public Response createLike(@RequestBody Like like){
+    public Response createLike(@RequestBody Like like, @RequestHeader Map<String, String> headers){
+        //Verify the JWT - Andrew
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
         Response response;
         Like tempLike = this.likeService.createLike(like);
         if(tempLike != null){
@@ -36,8 +48,14 @@ public class LikeController {
     //Get all Likes by PostID
     //Angular send message we receive
     @GetMapping("{postId}")
-    public Response getLikeByPostId(@PathVariable Integer postId){
-    Response response;
+    public Response getLikeByPostId(@PathVariable Integer postId, @RequestHeader Map<String, String> headers){
+        //Verify the JWT - Andrew
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
+        Response response;
     List<Like> like = this.likeService.getLikeByPostId(postId);
         if(like != null){
         response = new Response(true, "Here is the likes of this post", like);
@@ -50,8 +68,15 @@ public class LikeController {
     //Get like PostId and by UserId
     //Angular send message we receive
     @GetMapping("{postId}/{userId}")
-    public Response getLikeByPostIdAndUserID(@PathVariable Integer postId, @PathVariable Integer userId){
-       Response response;
+    public Response getLikeByPostIdAndUserID(@PathVariable Integer postId, @PathVariable Integer userId,
+                                             @RequestHeader Map<String, String> headers){
+        //Verify the JWT - Andrew
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
+        Response response;
         boolean theyLikedIt = false;
        List<Like> like = this.likeService.getLikeByPostId(postId);
        Integer likeId = 0;
@@ -72,7 +97,13 @@ public class LikeController {
     //Delete a Like
     //Angular send message we receive
     @DeleteMapping("{likeId}")
-    public Response deleteLike(@PathVariable Integer likeId){
+    public Response deleteLike(@PathVariable Integer likeId, @RequestHeader Map<String, String> headers){
+        //Verify the JWT - Andrew
+        DecodedJWT decoded = jwtUtility.verify(headers.get("jwt"));
+        if(decoded == null){
+            return new Response(false, "Invalid token", null);
+        }
+
         Response response;
         Boolean deleteLike = this.likeService.deleteLike(likeId);
         if(deleteLike){
